Changed the way Samples interact with Gizmos, also simplified the Gizmos system
